Move test, add test for max
